Remove credentials plugin to avoid dependency confilct
Problem:
credentials-binding has dependency of "SSH Credentials" plugin. On installation it implicitly installs the latest "SSH Credentials" plugin as the dependency, which requires a newer v2.2.0 credentials plugin, so the claimed plugin version is not satisfied.

Solution:
Remove credentials plugin from the list, a proper version of it will be installed by dependencies.